// types/supabase.ts
export type Database = {
  public: {
    Tables: {
      Events: {
        Row: {
          created_at: string;
          date: string | null;
          description: string | null;
          id: number;
          image: string | null;
          link: string | null;
          title: string | null;
        };
        Insert: {
          created_at?: string;
          date?: string | null;
          description?: string | null;
          id?: number;
          image?: string | null;
          link?: string | null;
          title?: string | null;
        };
        Update: {
          created_at?: string;
          date?: string | null;
          description?: string | null;
          id?: number;
          image?: string | null;
          link?: string | null;
          title?: string | null;
        };
        Relationships: [];
      };
      News: {
        Row: {
          created_at: string;
          date: string | null;
          description: string | null;
          id: number;
          image: string | null;
          link: string | null;
          title: string | null;
        };
        Insert: {
          created_at?: string;
          date?: string | null;
          description?: string | null;
          id?: number;
          image?: string | null;
          link?: string | null;
          title?: string | null;
        };
        Update: {
          created_at?: string;
          date?: string | null;
          description?: string | null;
          id?: number;
          image?: string | null;
          link?: string | null;
          title?: string | null;
        };
        Relationships: [];
      };
      Promotions: {
        Row: {
          created_at: string;
          description: string | null;
          endDate: string | null;
          id: number;
          image: string | null;
          link: string | null;
          startDate: string | null;
          title: string | null;
        };
        Insert: {
          created_at?: string;
          description?: string | null;
          endDate?: string | null;
          id?: number;
          image?: string | null;
          link?: string | null;
          startDate?: string | null;
          title?: string | null;
        };
        Update: {
          created_at?: string;
          description?: string | null;
          endDate?: string | null;
          id?: number;
          image?: string | null;
          link?: string | null;
          startDate?: string | null;
          title?: string | null;
        };
        Relationships: [];
      };
      Talks: {
        Row: {
          date: string | null;
          description: string | null;
          id: number;
          img_date: string | null;
          img_description: string | null;
          img_title: string | null;
          title: string | null;
        };
        Insert: {
          date?: string | null;
          description?: string | null;
          id?: number;
          img_date?: string | null;
          img_description?: string | null;
          img_title?: string | null;
          title?: string | null;
        };
        Update: {
          date?: string | null;
          description?: string | null;
          id?: number;
          img_date?: string | null;
          img_description?: string | null;
          img_title?: string | null;
          title?: string | null;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
      PublicSchema["Views"])
  ? (PublicSchema["Tables"] &
      PublicSchema["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R;
    }
    ? R
    : never
  : never;

export type TablesInsert<
  PublicTableNameOrOptions extends
    | keyof PublicSchema["Tables"]
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicTableNameOrOptions["schema"]]["Tables"]
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? Database[PublicTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I;
    }
    ? I
    : never
  : PublicTableNameOrOptions extends keyof PublicSchema["Tables"]
  ? PublicSchema["Tables"][PublicTableNameOrOptions] extends {
      Insert: infer I;
    }
    ? I
    : never
  : never;

export type TablesUpdate<
  PublicTableNameOrOptions extends
    | keyof PublicSchema["Tables"]
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicTableNameOrOptions["schema"]]["Tables"]
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? Database[PublicTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U;
    }
    ? U
    : never
  : PublicTableNameOrOptions extends keyof PublicSchema["Tables"]
  ? PublicSchema["Tables"][PublicTableNameOrOptions] extends {
      Update: infer U;
    }
    ? U
    : never
  : never;

export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
  ? PublicSchema["Enums"][PublicEnumNameOrOptions]
  : never;